spotify_rec_system/requirements.txt
Flask
spotipy
pandas
numpy
scikit-learn
torch
python-dotenv
requests
kafka-python
redis
